requirements.txt
google-cloud-compute==1.19.2
google-cloud-container==2.40.0
google-cloud-storage==2.16.0
google-api-python-client==2.118.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-resource-manager==1.12.3
pyVmomi==8.0.3.0.1
openpyxl==3.1.2
PyYAML==6.0.1
python-dotenv==1.0.1
